// frontend/src/client/types.gen.ts
export type ProductCreate = {
    name: string;
    description?: (string | null);
    price: number;
    quantity: number;
    category: string;
};